requirements.txt
pdfplumber
crewai
crewai-tools
google-genai
python-dotenv